+++ server-plugin/Code/Systems/ConfigManager.h
@@ -25,7 +25,7 @@ public:
 	float m_smoke_timetobang;
 	float m_smoke_time;
 	static float tickinterval;
-	int tickrate_override;
+	float tickrate_override;
 
 	// virtual functions
 	int vfid_getdatadescmap;
